requirements.txt
amqp==2.5.0
billiard==3.6.0.0
boto3==1.9.185
botocore==1.12.185
celery==4.3.0
Django==2.2.3
django-adminlte2==0.4.1
django-mysql==3.2.0
django-split-settings==0.3.0
docutils==0.14
jmespath==0.9.4
kombu==4.6.3
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
